Move `RemoteFeedItem` and `LocalFeedItem` to their own files
