+++ MyEssentialFeed/MyEssentialFeed/Feed API/FeedItemsMapper.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-struct RemoteFeedItem: Decodable {
-    let id: UUID
-    let description: String?
-    let location: String?
-    let image: URL
-}
-
 internal final class FeedItemMapper {
     private struct Root: Decodable {
         let items: [RemoteFeedItem]
